backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
python-dotenv
sqlalchemy
databases
google-generativeai